feat(frontend): integrate onboarding tour into root and dashboard layouts
