+++ apps/frontend/app/root.tsx
@@ -31,6 +31,7 @@ import {
 	useNavigation,
 } from "react-router";
 import "mantine-datatable/styles.layer.css";
+import { OnboardingTour } from "@gfazioli/mantine-onboarding-tour";
 import { Toaster } from "~/components/toaster";
 import { LOGO_IMAGE_URL, queryClient } from "~/lib/generals";
 import {
@@ -38,6 +39,7 @@ import {
 	extendResponseHeaders,
 	getToast,
 } from "~/lib/utilities.server";
+import { tourSteps, useOnboardingTour } from "./lib/state/general";
 
 const theme = createTheme({
 	fontFamily: "Poppins",
@@ -106,6 +108,9 @@ export const loader = async ({ request }: LoaderFunctionArgs) => {
 export default function App() {
 	const navigation = useNavigation();
 	const loaderData = useLoaderData<typeof loader>();
+	const { isTourStarted, setIsTourStarted } = useOnboardingTour();
+
+	const stopTour = () => setIsTourStarted(false);
 
 	return (
 		<html lang="en">
@@ -122,33 +127,40 @@ export default function App() {
 				<ColorSchemeScript forceColorScheme={loaderData.defaultColorScheme} />
 			</head>
 			<body>
-				<QueryClientProvider client={queryClient}>
-					<MantineProvider
-						theme={theme}
-						classNamesPrefix="mnt"
-						forceColorScheme={loaderData.defaultColorScheme}
-					>
-						<ModalsProvider>
-							{["loading", "submitting"].includes(navigation.state) ? (
-								<Loader
-									top={10}
-									size="sm"
-									right={10}
-									pos="fixed"
-									color="yellow"
-									style={{ zIndex: 10 }}
-								/>
-							) : null}
-							<Toaster toast={loaderData.toast} />
-							<Flex style={{ flexGrow: 1 }} mih="100vh">
-								<Outlet />
-							</Flex>
-							<ScrollRestoration />
-							<Scripts />
-						</ModalsProvider>
-					</MantineProvider>
-					<ReactQueryDevtools buttonPosition="top-right" />
-				</QueryClientProvider>
+				<OnboardingTour
+					tour={tourSteps}
+					started={isTourStarted}
+					onOnboardingTourEnd={stopTour}
+					onOnboardingTourClose={stopTour}
+				>
+					<QueryClientProvider client={queryClient}>
+						<MantineProvider
+							theme={theme}
+							classNamesPrefix="mnt"
+							forceColorScheme={loaderData.defaultColorScheme}
+						>
+							<ModalsProvider>
+								{["loading", "submitting"].includes(navigation.state) ? (
+									<Loader
+										top={10}
+										size="sm"
+										right={10}
+										pos="fixed"
+										color="yellow"
+										style={{ zIndex: 10 }}
+									/>
+								) : null}
+								<Toaster toast={loaderData.toast} />
+								<Flex style={{ flexGrow: 1 }} mih="100vh">
+									<Outlet />
+								</Flex>
+								<ScrollRestoration />
+								<Scripts />
+							</ModalsProvider>
+						</MantineProvider>
+						<ReactQueryDevtools buttonPosition="top-right" />
+					</QueryClientProvider>
+				</OnboardingTour>
 			</body>
 		</html>
 	);
